+++ src/main/java/pl/plajer/buildbattle/menus/particles/ParticleItem.java
@@ -1,0 +1,72 @@
+/*
+ * BuildBattle 4 - Ultimate building competition minigame
+ * Copyright (C) 2018  Plajer's Lair - maintained by Plajer and Tigerpanzer
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package pl.plajer.buildbattle.menus.particles;
+
+import org.bukkit.Particle;
+import org.bukkit.inventory.ItemStack;
+
+import pl.plajer.buildbattle.handlers.ChatManager;
+
+/**
+ * Created by Tom on 23/08/2015.
+ */
+public class ParticleItem {
+
+  private ItemStack itemStack;
+  private Particle effect;
+  private String permission;
+  private boolean enabled = true;
+
+  public String getPermission() {
+    return permission;
+  }
+
+  public void setPermission(String permission) {
+    this.permission = permission;
+  }
+
+  public boolean isEnabled() {
+    return enabled;
+  }
+
+  public void setEnabled(boolean enabled) {
+    this.enabled = enabled;
+  }
+
+  public String getDisplayName() {
+    return ChatManager.colorRawMessage(itemStack.getItemMeta().getDisplayName());
+  }
+
+  public Particle getEffect() {
+    return effect;
+  }
+
+  public void setEffect(Particle effect) {
+    this.effect = effect;
+  }
+
+  public ItemStack getItemStack() {
+    return itemStack;
+  }
+
+  public void setItemStack(ItemStack itemStack) {
+    this.itemStack = itemStack;
+  }
+
+}